astra/requirements.txt
Django==5.2.6
celery==5.5.3
redis==5.2.1
requests==2.32.3
gunicorn==22.0.0
dj-database-url==3.0.1
psycopg2-binary==2.9.10
whitenoise==6.9.0
google-generativeai==0.8.5
python-dotenv==1.0.1
django_celery_results==2.6.0
Markdown==3.6
daphne
uvicorn
